Move config and templates to the `cmd` directory
